requirements.txt
asgiref==3.3.1
certifi==2020.12.5
click==7.1.2
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.2
django-extensions==3.1.1
django-filter==2.4.0
django-heroku==0.3.1
django-taggit==1.3.0
djangorestframework==3.12.3
fake-factory==9999.9.9
gunicorn==20.0.4
h11==0.12.0
Markdown==3.3.4
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.67
six==1.15.0
sqlparse==0.4.1
uvicorn==0.13.4
whitenoise==5.2.0
wincertstore==0.2
